Reads the line numbers of project
NOTE : Instead of running it. Use terminal to run it ( for example : python Lines.py __aFileName__.py ) . If you put this project in a folder ( for example : Python Filesss/s4/Lines.py ), cd that folder you've putted the project in it ( for example : cd s4 ) and then run it on terminal.